+++ components/auth/Login/index.tsx
@@ -13,12 +13,35 @@ const LoginPage = ({
   onPasswordChange,
 }: LoginPageProps) => (
   <React.Fragment>
-    <h3>Login</h3>
-    <LoginForm
-      onLogin={onLogin}
-      onEmailChange={onEmailChange}
-      onPasswordChange={onPasswordChange}
-    />
+    <div className="loginContainer">
+      <h3>Login</h3>
+      <LoginForm
+        onLogin={onLogin}
+        onEmailChange={onEmailChange}
+        onPasswordChange={onPasswordChange}
+      />
+    </div>
+    <style jsx>{`
+      h3 {
+        text-align: center;
+      }
+      .loginContainer {
+        width: 50%;
+        margin: auto;
+      }
+
+      @media only screen and (max-width: 800px) {
+        .loginContainer {
+          width: 75%;
+        }
+      }
+
+      @media only screen and (max-width: 600px) {
+        .loginContainer {
+          width: 100%;
+        }
+      }
+    `}</style>
   </React.Fragment>
 );