+++ store/user/actions.ts
@@ -1,4 +1,5 @@
 import * as constants from './consts';
+import Router from 'next/router';
 import {UserState, UserLogin} from './reducers';
 import get from 'lodash.get';
 import * as moment from 'moment';
@@ -50,6 +51,13 @@ export const completedLogin = (user: UserState): CompletedLogin => {
   };
 };
 
+export const logoutUser = () => {
+  return (dispatch: any) => {
+    Router.push('/');
+    dispatch(removeUser());
+  };
+};
+
 export const loginUser = (userLogin: UserLogin) => {
   return (dispatch: any) => {
     dispatch(startingLogin());
@@ -70,6 +78,7 @@ export const loginUser = (userLogin: UserLogin) => {
           isActive: true,
         };
         dispatch(completedLogin(user));
+        Router.push('/admin');
       },
       error: error => {
         console.error(